flask-example/flask-mysql/flask-dance-mysql/requirements.txt
flask 
flask-dance 
flask-sqlalchemy 
flask-login 
pymysql 
python-dotenv